refactor(csrc): extract memory models and sim runner helpers
